Added panels to show with player is active

+++ Assets/Scripts/Blocks/BlocksSelect.cs
@@ -77,7 +77,7 @@ namespace Harmonies.Blocks
 
         private void OnMouseDown()
         {
-            if (_turnManager.IndexActualPlayer != (int)NetworkManager.Singleton.LocalClientId)
+            if (_turnManager.IndexActualPlayer != NetworkManager.Singleton.LocalClientId)
                 return;
 
             if (!_isActive)
